trying to run train2 rn
