+++ requirements.txt
@@ -1,0 +1,61 @@
+absl-py==2.1.0
+aiohappyeyeballs==2.4.3
+aiohttp==3.11.6
+aiosignal==1.3.1
+attrs==24.2.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+chex==0.1.87
+e3nn==0.5.4
+e3nn-jax==0.20.7
+e3x==1.0.2
+etils==1.10.0
+filelock==3.16.1
+flax==0.10.2
+frozenlist==1.5.0
+fsspec==2024.10.0
+humanize==4.11.0
+idna==3.10
+importlib_resources==6.4.5
+jax==0.4.35
+jaxlib==0.4.35
+jaxtyping==0.2.36
+Jinja2==3.1.4
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+mdurl==0.1.2
+ml_dtypes==0.5.0
+more-itertools==10.5.0
+mpmath==1.3.0
+msgpack==1.1.0
+multidict==6.1.0
+nest-asyncio==1.6.0
+networkx==3.4.2
+numpy==2.1.3
+opt-einsum-fx==0.1.4
+opt_einsum==3.4.0
+optax==0.2.4
+orbax-checkpoint==0.9.1
+packaging==24.2
+propcache==0.2.0
+protobuf==5.28.3
+psutil==6.1.0
+Pygments==2.18.0
+pyparsing==3.2.0
+PyYAML==6.0.2
+requests==2.32.3
+rich==13.9.4
+scipy==1.14.1
+sympy==1.13.1
+tensorstore==0.1.68
+toolz==1.0.0
+torch==2.5.1
+torch-geometric==2.6.1
+torch_scatter @ git+https://github.com/curtischong/pytorch_scatter.git@91db04d32d5e9514a3fa76fcc60251cab09bcf45
+torchtyping==0.1.5
+tqdm==4.67.0
+typeguard==2.13.3
+typing_extensions==4.12.2
+urllib3==2.2.3
+yarl==1.17.2
+zipp==3.21.0
